+++ moor/CHANGELOG.md
@@ -6,7 +6,9 @@
 - Add the `scoped_dart_components` builder option to generate a function for
   [Dart components in SQL](https://moor.simonbinder.eu/docs/using-sql/moor_files/#dart-components-in-sql).
   The function's parameters reflect tables that are in scope.
-
+- Add the `closeUnderlyingOnClose` argument to `VmDatabase.opened` as an option
+  to keep the underlying instance opened when closing the `VmDatabase`.
+ 
 ## 4.3.2
 
 - Fix `insertReturning` not updating streams